pyproject.toml
[project]
name = "relouderul"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "watchfiles>=1.1.1",
]
